Separated out CLIPWrapper to own file
